Move the test suite onto cents, bytes32 ids and payment references
The ETH figure each purchase test was written with stays the ETH argument.
The recorded price becomes a cents value, and the ceiling tests were the
ones that had to change meaning rather than shape: they bound what is
recorded now, not what is sent.

Every purchase draws a fresh payment reference. A reference is spendable
once protocol-wide, so a test buying twice would otherwise fail on replay
instead of on what it was written to check.

+++ test/foundry/unit-testing/upgrade/StorageLayout.t.sol
@@ -108,7 +108,7 @@ contract StorageLayoutTest is DeployerBase {
         // One entry long, with a price written straight into the element the length implies
         vm.store(_target, bytes32(uint256(3)), bytes32(uint256(1)));
         vm.store(_target, bytes32(uint256(keccak256(abi.encode(uint256(3)))) + 1), bytes32(uint256(0xB1)));
-        (, uint256 price) = _wallet.transactionHistory(0);
+        (, uint64 price,) = _wallet.transactionHistory(0);
         assertEq(price, 0xB1, "ESIMWallet.transactionHistory must read slot 3");
 
         vm.store(_target, bytes32(uint256(4)), bytes32(uint256(uint160(SENTINEL))));
